Feature/node postgres db support (#1402)
* Updated migrations and sequelize configuration

* Merge branch 'feature/ha-deployment' into feature/node-postgres-db-support

* Package.json version updated

* Updated install postgres script

* Updated install postgres script

* Updated docker file to install postgres

* Added missing library for postgres

* Updated testnet dockerfile

* Updatet listen addresses for postgres configuration

* Using psql 12

* Added psql start every time psql is used for the first time

* Docker config changes

* Updated migrations

* Added create user root so we can acces db

* Update setup.js

* Update setup.js

* Update migrations and setup.js

* Update docker file

* Refactor setup.js

* Refactor sql migration

* Update register-node

+++ migrations/202010171600125-update-handler-ids-add-timestamp.js
@@ -4,11 +4,11 @@ module.exports = {
             'handler_ids',
             'timestamp',
             {
-                type: Sequelize.INTEGER,
+                type: Sequelize.BIGINT,
             },
         );
         const now = Date.now();
-        await queryInterface.sequelize.query(`UPDATE handler_ids SET timestamp = ${now}`);
+        await queryInterface.sequelize.query(`UPDATE handler_ids SET timestamp = ${now} `);
     },
     down: async (queryInterface) => {
         await queryInterface.removeColumn('handler_ids', 'timestamp');
